requirements.txt
# Core
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
# SQLite is used as the default database for simplicity
# psycopg2 is not needed for SQLite

# Authentication
Flask-JWT-Extended==4.5.2
Flask-Session==0.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# API Documentation
flasgger==0.9.7.1
pyyaml==6.0.1

# AI/ML
openai==1.3.5
whisper==1.0.0

# File Processing
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Utilities
python-dateutil==2.8.2
python-slugify==8.0.1
email-validator==2.0.0

# Development
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0
